+++ src/main/java/com/example/mind_laundromat/cbt/repository/FeedbackRepository.java
@@ -3,5 +3,5 @@ package com.example.mind_laundromat.cbt.repository;
 import com.example.mind_laundromat.cbt.entity.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
+public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
 }
